Fixes/improvements
