+++ src/lib/roboChimp.ts
@@ -1,14 +1,13 @@
-import { PerkTier, formatOrdinal } from '@oldschoolgg/toolkit';
+import { formatOrdinal } from '@oldschoolgg/toolkit';
 import type { TriviaQuestion, User } from '@prisma/robochimp';
 import { calcWhatPercent, round, sumArr } from 'e';
 import deepEqual from 'fast-deep-equal';
-import { pick } from 'lodash';
 import type { Bank } from 'oldschooljs';
 
-import { SupportServer } from '../config';
-import { BOT_TYPE, BitField, Roles, globalConfig, masteryKey } from './constants';
+import { BOT_TYPE, globalConfig, masteryKey } from './constants';
 import { getTotalCl } from './data/Collections';
 import { calculateMastery } from './mastery';
+import { cacheRoboChimpUser } from './perkTier';
 import { MUserStats } from './structures/MUserStats';
 
 export type RobochimpUser = User;
@@ -117,78 +116,4 @@ WHERE osb_cl_percent >= (SELECT osb_cl_percent FROM public.user WHERE id = ${Big
 	)[0].count;
 
 	return formatOrdinal(clPercentRank);
-}
-
-const robochimpCachedKeys = [
-	'bits',
-	'github_id',
-	'patreon_id',
-	'perk_tier',
-	'main_account',
-	'ironman_alts',
-	'premium_balance_expiry_date',
-	'premium_balance_tier'
-] as const;
-type CachedRoboChimpUser = Pick<User, (typeof robochimpCachedKeys)[number]>;
-
-export const roboChimpCache = new Map<string, CachedRoboChimpUser>();
-
-export async function populateRoboChimpCache() {
-	const users = await roboChimpClient.user.findMany({
-		select: {
-			id: true,
-			bits: true,
-			github_id: true,
-			patreon_id: true,
-			perk_tier: true,
-			main_account: true,
-			ironman_alts: true,
-			premium_balance_expiry_date: true,
-			premium_balance_tier: true
-		},
-		where: {
-			perk_tier: {
-				not: 0
-			}
-		}
-	});
-	for (const user of users) {
-		roboChimpCache.set(user.id.toString(), user);
-	}
-	debugLog(`Populated RoboChimp cache with ${users.length} users.`);
-}
-
-function cacheRoboChimpUser(user: RobochimpUser) {
-	if (user.perk_tier === 0) return;
-	roboChimpCache.set(user.id.toString(), pick(user, robochimpCachedKeys));
-}
-
-export function getPerkTierSync(user: string | MUser) {
-	const elligibleTiers = [];
-	if (typeof user !== 'string') {
-		if (
-			[BitField.isContributor, BitField.isModerator, BitField.IsWikiContributor].some(bit =>
-				user.bitfield.includes(bit)
-			)
-		) {
-			return PerkTier.Four;
-		}
-
-		if (
-			user.bitfield.includes(BitField.IsPatronTier1) ||
-			user.bitfield.includes(BitField.HasPermanentTierOne) ||
-			user.bitfield.includes(BitField.BothBotsMaxedFreeTierOnePerks)
-		) {
-			elligibleTiers.push(PerkTier.Two);
-		} else {
-			const guild = globalClient.guilds.cache.get(SupportServer);
-			const member = guild?.members.cache.get(user.id);
-			if (member && [Roles.Booster].some(roleID => member.roles.cache.has(roleID))) {
-				elligibleTiers.push(PerkTier.One);
-			}
-		}
-	}
-
-	elligibleTiers.push(roboChimpCache.get(typeof user === 'string' ? user : user.id)?.perk_tier ?? 0);
-	return Math.max(...elligibleTiers);
 }
