+++ src/mahoji/lib/inhibitors.ts
@@ -5,7 +5,7 @@ import { ComponentType, PermissionsBitField } from 'discord.js';
 import { OWNER_IDS, SupportServer } from '../../config';
 import { BLACKLISTED_GUILDS, BLACKLISTED_USERS } from '../../lib/blacklists';
 import { BadgesEnum, BitField, Channel, DISABLED_COMMANDS, minionBuyButton } from '../../lib/constants';
-import { getPerkTierSync } from '../../lib/roboChimp';
+import { getPerkTierSync } from '../../lib/perkTier';
 import type { CategoryFlag } from '../../lib/types';
 import { minionIsBusy } from '../../lib/util/minionIsBusy';
 import { mahojiGuildSettingsFetch, untrustedGuildSettingsCache } from '../guildSettings';
