+++ src/tasks/minions/collectingActivity.ts
@@ -5,7 +5,7 @@ import { MorytaniaDiary, userhasDiaryTier } from '../../lib/diaries';
 import type { CollectingOptions } from '../../lib/types/minions';
 import { handleTripFinish } from '../../lib/util/handleTripFinish';
 import { updateBankSetting } from '../../lib/util/updateBankSetting';
-import { collectables } from '../../mahoji/lib/abstracted_commands/collectCommand';
+import { collectables } from '../../mahoji/lib/collectables';
 
 export const collectingTask: MinionTask = {
 	type: 'Collecting',
